+++ src/utils.hpp
@@ -1,5 +1,5 @@
 /**
- * Used for qsort on doubles. If d1 < d2, a number < 0 returned
+ * Used for qsort on floats. If d1 < d2, a number < 0 returned
  * if d1 > d2, a number > 0 returned
  * 
 */
@@ -8,9 +8,9 @@ int dblCmp(const void * d1, const void * d2);
 /**
  * Convert the ppm value of a given mass to a value in daltons
  * 
- * @param mass      double  the mass to calculate the dalton tolerance for
+ * @param mass      float  the mass to calculate the dalton tolerance for
  * @param ppmTol    int     the tolerance in ppm
  * 
- * @return double   the tolerance in daltons 
+ * @return float   the tolerance in daltons 
 */
-double ppmToDa(double mass, int ppmTol);
+float ppmToDa(float mass, int ppmTol);
